pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "decky"
version = "0.2.0"
description = "A modular YAML-driven Stream Deck controller for Linux"
readme = "README.md"
requires-python = ">=3.8"
authors = [
    {name = "Austin Nicholas", email = "anicholas@netflix.com"}
]
dependencies = [
    "streamdeck>=0.9.0",
    "pillow>=9.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "flake8-docstrings>=1.7.0",
    "flake8-bugbear>=23.0.0",
    "mypy>=1.0.0",
    "isort>=5.12.0",
    "bandit[toml]>=1.7.0",
    "safety>=2.3.0",
    "pre-commit>=3.0.0",
]
dbus = [
    "dbus-python>=1.2.0",
    "PyGObject>=3.40.0",
]

[project.scripts]
decky = "decky.main:main"
deckyctl = "decky.cli:main"

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
pythonpath = ["src"]
addopts = [
    "--verbose",
    "--cov=decky",
    "--cov-report=term-missing",
    "--cov-report=html",
]

[tool.black]
line-length = 100
target-version = ["py38", "py39", "py310", "py311"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.coverage.run]
source = ["src/decky"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/venv/*",
    "*/.venv/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
    "pass",
    "except ImportError:",
]

[tool.isort]
profile = "black"
line_length = 100
known_first_party = ["decky"]
skip_gitignore = true

[tool.bandit]
exclude_dirs = [
    "/tests",
    "/.venv",
    "/venv",
]
skips = ["B404", "B603", "B607"]

[tool.flake8]
max-line-length = 100
extend-ignore = ["E203", "W503"]
exclude = [
    ".git",
    "__pycache__",
    ".venv",
    "venv",
    "build",
    "dist",
]